+++ qt_ui/windows/newgame/QNewGameWizard.py
@@ -150,6 +150,7 @@ class NewGameWizard(QtWidgets.QWizard):
         logging.info("New campaign theater: %s", theater.terrain.name)
 
         settings = self.settings_page.settings_widget.settings
+        settings.apply_mod_cloud_presets()
 
         generator = GameGenerator(
             blue_faction,